Create team stats module and test files
